// venminder/control-assessments/src/shared-from-dcdev/resources/features/vm-date-picker/vm-date-picker.enums.ts
export enum VmDatePickerDisplayType {
  PredefinedRangesOnly = 1,
  DateViewOnly = 2,
  Full = 4,
}

export enum VmDatePickerCSSDefinitions {
  Selected = 'selected',
  Start = 'start',
  End = 'end',
  Lighter = 'lighter',
  DayNotInMonth = 'day-not-in-month',
  Today = 'today',
}

export enum VmDatePickerDateFormats {
  DateFnsSlash = 'MM/dd/yyyy',
}

export enum VmDatePickerSearchTypes {
  SearchDateRange = 'search-date-range',
}